requirements.txt
yfinance>=0.2.40
jinja2>=3.1.3
anthropic>=0.40.0
feedparser>=6.0.10
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
